Revert "Move variant style handling for Infocard to seperate file and use some of it for TabbedCard"
This reverts commit 94fa82b741eea20f1935ac6ec8bd7de885bc8f50.
